Refactorización de código según estándares
