+++ package.json
@@ -1,25 +1,27 @@
 {
-    "name": "react-crud",
+    "name": "ebar",
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@material-ui/core": "^4.11.3",
+        "@material-ui/icons": "^4.11.2",
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^9.3.2",
         "@testing-library/user-event": "^7.1.2",
         "axios": "^0.21.1",
         "babel-core": "^6.26.3",
-        "bootstrap": "^4.4.1",
         "html-webpack-plugin": "^5.3.1",
         "jquery": "^3.6.0",
+        "notistack": "^1.0.5",
         "popper.js": "^1.16.1",
         "react": "^16.13.0",
         "react-dom": "^16.13.0",
         "react-router-dom": "^5.1.2",
         "react-scripts": "^4.0.3",
         "react-validation": "^3.0.7",
-        "webpack": "4.44.2",
+        "typescript": "^4.2.3",
         "validator": "^13.5.2",
-        "typescript": "^4.2.3"
+        "webpack": "4.44.2"
     },
     "scripts": {
         "start": "react-scripts start",
