+++ src/components/bar-list.component.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import BarDataService from "../services/bar.service";
+import BarDataService from "../../services/bar.service";
 
 export default class BarList extends Component {
