+++ src/components/bar.component.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import BarDataService from "../services/bar.service";
+import BarDataService from "../../services/bar.service";
 
 export default class Bar extends Component{
     constructor(props) {
